refactor: context api 변경으로 인한 컴포넌트의 hook 생성
