+++ src/components/ProductControl/ProductControl.tsx
@@ -1,36 +1,11 @@
 import * as S from './ProductControl.styled';
 import Select from '../common/Select/Select';
-import { productApi } from '../../api/productApi';
 import { CategoryOptions, SortOptions } from '../../constants/selectOptions';
-import { useState } from 'react';
-import { CategoryOptionType, ProductControlProps, SelectType, SortOptionType } from './types';
+import { CategoryOptionType, ProductControlProps, SortOptionType } from './types';
+import { useProductControl } from './hooks/useProductControl';
 
-function ProductControl({ setProductList, setErrorMessage }: ProductControlProps) {
-  const [category, setCategory] = useState<CategoryOptionType>('');
-  const [sort, setSort] = useState<SortOptionType>('price,asc');
-
-  async function handleSelectChange(selectedValue: CategoryOptionType | SortOptionType, type: SelectType) {
-    try {
-      const newCategory = type === 'category' ? selectedValue : category;
-      const newSort = type === 'sort' ? selectedValue : sort;
-
-      if (type === 'category') {
-        setCategory(selectedValue as CategoryOptionType);
-      } else {
-        setSort(selectedValue as SortOptionType);
-      }
-
-      const rawProductList = await productApi.get({
-        category: newCategory,
-        sort: newSort,
-      });
-      setProductList(rawProductList);
-    } catch (error) {
-      if (error instanceof Error) {
-        setErrorMessage(error.message);
-      }
-    }
-  }
+function ProductControl({ setProductList }: ProductControlProps) {
+  const { handleSelectChange } = useProductControl(setProductList);
 
   return (
     <S.ProductControlContainer>
