+++ src/components/ProductControl/types/index.ts
@@ -3,7 +3,6 @@ import { ResponseProduct } from '../../../api/types';
 
 export interface ProductControlProps {
   setProductList: Dispatch<SetStateAction<ResponseProduct[]>>;
-  setErrorMessage: Dispatch<SetStateAction<string>>;
 }
 
 export type CategoryOptionType = '' | '식료품' | '패션잡화';
