+++ src/components/ProductList/ProductList.tsx
@@ -1,13 +1,12 @@
 import * as S from './ProductList.styled';
 import ProductItem from '../ProductItem/ProductItem';
 import { ResponseProduct } from '../../api/types';
-import { Dispatch } from 'react';
 
-function ProductList({ productList, setErrorMessage }: { productList: ResponseProduct[]; setErrorMessage: Dispatch<React.SetStateAction<string>> }) {
+function ProductList({ productList }: { productList: ResponseProduct[] }) {
   return (
     <S.ProductListContainer>
       {productList.map((product) => (
-        <ProductItem key={product.id} product={product} setErrorMessage={setErrorMessage} />
+        <ProductItem key={product.id} product={product} />
       ))}
     </S.ProductListContainer>
   );
