+++ src/pages/productListPage/context/useCartContext.tsx
@@ -4,14 +4,22 @@ import { ResponseCartItem } from '../../../api/types';
 type CartContextType = {
   cartItemList: ResponseCartItem[];
   setCartItemList: React.Dispatch<React.SetStateAction<ResponseCartItem[]>>;
+  errorMessage: string;
+  setErrorMessage: React.Dispatch<React.SetStateAction<string>>;
+  isLoading: boolean;
+  setIsLoading: React.Dispatch<React.SetStateAction<boolean>>;
 };
 
 export const CartContext = createContext<CartContextType | null>(null);
 
 export const CartProvider = ({ children }: PropsWithChildren) => {
   const [cartItemList, setCartItemList] = useState<ResponseCartItem[]>([]);
+  const [errorMessage, setErrorMessage] = useState('');
+  const [isLoading, setIsLoading] = useState(true);
 
-  return <CartContext.Provider value={{ cartItemList, setCartItemList }}>{children}</CartContext.Provider>;
+  return (
+    <CartContext.Provider value={{ cartItemList, setCartItemList, errorMessage, setErrorMessage, isLoading, setIsLoading }}>{children}</CartContext.Provider>
+  );
 };
 
 export const useCartItemList = () => {
